# api/openapi.yaml
openapi: '3.0.3'
info:
  title: 'Motocorp Dealership Management System API'
  description: |
    A comprehensive cloud-based dealership management system for motorcycle dealerships.
    This API provides multi-tenant architecture with complete CRUD operations for:
    - Inventory Management (motorcycles, parts, accessories)
    - Customer Relationship Management (CRM)
    - Sales Pipeline Management
    - Service Appointment Management
    - User Management with role-based access control
    - Real-time notifications and activity feeds
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`
    
    ## Multi-Tenancy
    The API supports multi-tenant architecture where each organization has its own isolated data.
    The tenant context is automatically resolved from the authenticated user's JWT token.
    
    ## Rate Limiting
    API requests are rate-limited to ensure fair usage and system stability.
    
    ## Error Handling
    Standard HTTP status codes are used. Error responses include detailed error messages
    and validation errors where applicable.
  version: '1.0.0'
  contact:
    name: 'Motocorp DMS Support'
    email: 'support@motocorpdms.com'
    url: 'https://motocorpdms.com/support'
  license:
    name: 'Proprietary'
servers:
  - url: 'https://api.motocorpdms.com/v1'
    description: 'Production server'
...
    description: 'Staging server'
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Retrieves a paginated list of users within the organization
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: role
          in: query
          description: Filter users by role
          schema:
            type: string
            enum: [admin, manager, sales, service, finance]
        - name: dealershipId
          in: query
          description: Filter users by dealership ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'

    post:
      tags:
        - Users
      summary: Create a new user
      description: Creates a new user within the organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieves a specific user by their ID
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Users
      summary: Update user
      description: Updates an existing user's information
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Soft deletes a user (sets isActive to false)
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Customer Management
...
  /sales:
    get:
      tags:
        - Sales
      summary: List sales
      description: Retrieves a paginated list of sales with filtering options
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: stage
          in: query
          description: Filter by sales stage
          schema:
            type: string
            enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
        - name: salesPersonId
          in: query
          description: Filter by sales person
          schema:
            type: string
            format: uuid
        - name: dealershipId
          in: query
          description: Filter by dealership
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sales retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSales'

    post:
      tags:
        - Sales
      summary: Create a new sale
      description: Creates a new sales opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleDto'
      responses:
        '201':
          description: Sale created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /sales/{id}:
    get:
      tags:
        - Sales
      summary: Get sale by ID
      description: Retrieves a specific sale with detailed information
      parameters:
        - name: id
          in: path
          required: true
          description: Sale ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sale retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleDetail'
        '404':
          description: Sale not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Sales
      summary: Update sale
      description: Updates an existing sale
      parameters:
        - name: id
          in: path
          required: true
          description: Sale ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSaleDto'
      responses:
        '200':
          description: Sale updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          description: Sale not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Sales
      summary: Delete sale
      description: Deletes a sales record
      parameters:
        - name: id
          in: path
          required: true
          description: Sale ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Sale deleted successfully
        '404':
          description: Sale not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sales/{id}/stage:
    patch:
      tags:
        - Sales
      summary: Update sales stage
      description: Updates the stage of a sales opportunity
      parameters:
        - name: id
          in: path
          required: true
          description: Sale ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSaleStageDto'
      responses:
        '200':
          description: Sales stage updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          description: Sale not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Service Appointments
  /service-appointments:
    get:
      tags:
        - Service Appointments
      summary: List service appointments
      description: Retrieves a paginated list of service appointments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by appointment status
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
        - name: technicianId
          in: query
          description: Filter by technician
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          description: Filter by appointment date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Service appointments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedServiceAppointments'

    post:
      tags:
        - Service Appointments
      summary: Create service appointment
      description: Creates a new service appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceAppointmentDto'
      responses:
        '201':
          description: Service appointment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAppointment'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /service-appointments/{id}:
    get:
      tags:
        - Service Appointments
      summary: Get service appointment by ID
      description: Retrieves a specific service appointment
      parameters:
        - name: id
          in: path
          required: true
          description: Service appointment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service appointment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAppointmentDetail'
        '404':
          description: Service appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Service Appointments
      summary: Update service appointment
      description: Updates an existing service appointment
      parameters:
        - name: id
          in: path
          required: true
          description: Service appointment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceAppointmentDto'
      responses:
        '200':
          description: Service appointment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAppointment'
        '404':
          description: Service appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Service Appointments
      summary: Cancel service appointment
      description: Cancels a service appointment
      parameters:
        - name: id
          in: path
          required: true
          description: Service appointment ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Service appointment cancelled successfully
        '404':
          description: Service appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Notifications
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Bearer token authentication. Format: Bearer <your_jwt_token>'
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UserIdParam:
      name: id
      in: path
      required: true
      description: User ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: '123e4567-e89b-12d3-a456-426614174000'
        organizationId:
          type: string
          format: uuid
          description: Organization ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        dealershipId:
          type: string
          format: uuid
          nullable: true
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        email:
          type: string
          format: email
          description: User's email address
          example: 'john.doe@example.com'
        firstName:
          type: string
          description: User's first name
          example: 'John'
        lastName:
          type: string
          description: User's last name
          example: 'Doe'
        role:
          type: string
          enum: [admin, manager, sales, service, finance]
          description: User's role
          example: 'sales'
        permissions:
          type: object
          nullable: true
          description: User's specific permissions
          example: {"canViewReports": true, "canEditInventory": false}
        isActive:
          type: boolean
          description: Whether the user is active
          example: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
          example: '2024-01-15T10:30:00Z'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T10:30:00Z'
    CreateUserDto:
      type: object
      required:
        - email
        - firstName
        - lastName
        - role
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: 'new.user@example.com'
        firstName:
          type: string
          minLength: 2
          description: User's first name
          example: 'Alice'
        lastName:
          type: string
          minLength: 2
          description: User's last name
          example: 'Johnson'
        dealershipId:
          type: string
          format: uuid
          nullable: true
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        role:
          type: string
          enum: [admin, manager, sales, service, finance]
          description: User's role
          example: 'manager'
        permissions:
          type: object
          nullable: true
          description: User's specific permissions
          example: {"canViewReports": true, "canEditInventory": true}
    UpdateUserDto:
      type: object
      properties:
        firstName:
          type: string
          minLength: 2
          description: User's first name
          example: 'Alice'
        lastName:
          type: string
          minLength: 2
          description: User's last name
          example: 'Johnson'
        dealershipId:
          type: string
          format: uuid
          nullable: true
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        role:
          type: string
          enum: [admin, manager, sales, service, finance]
          description: User's role
          example: 'manager'
        permissions:
          type: object
          nullable: true
          description: User's specific permissions
          example: {"canViewReports": true, "canEditInventory": true}
        isActive:
          type: boolean
          description: Whether the user is active
          example: true
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # Customer Management Schemas
...
    Sale:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique sale identifier
          example: '123e4567-e89b-12d3-a456-426614174000'
        customerId:
          type: string
          format: uuid
          description: Customer ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        dealershipId:
          type: string
          format: uuid
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        salesPersonId:
          type: string
          format: uuid
          nullable: true
          description: Sales person ID
          example: '123e4567-e89b-12d3-a456-426614174003'
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
          description: Sales stage
          example: 'proposal'
        totalAmount:
          type: number
          format: decimal
          nullable: true
          description: Total sale amount
          example: 16999.99
        items:
          type: array
          nullable: true
          items:
            type: object
            properties:
              inventoryId:
                type: string
                format: uuid
                example: '123e4567-e89b-12d3-a456-426614174004'
              quantity:
                type: integer
                example: 1
              unitPrice:
                type: number
                format: decimal
                example: 16999.99
              discount:
                type: number
                format: decimal
                nullable: true
                example: 500.00
          description: Sale items
        notes:
          type: string
          nullable: true
          description: Sale notes
          example: 'Customer interested in extended warranty'
        estimatedCloseDate:
          type: string
          format: date-time
          nullable: true
          description: Estimated close date
          example: '2024-02-15T00:00:00Z'
        actualCloseDate:
          type: string
          format: date-time
          nullable: true
          description: Actual close date
          example: '2024-02-10T14:30:00Z'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T10:30:00Z'
    SaleDetail:
      allOf:
        - $ref: '#/components/schemas/Sale'
        - type: object
          properties:
            customer:
              $ref: '#/components/schemas/Customer'
            salesPerson:
              $ref: '#/components/schemas/User'
              nullable: true
            itemsDetail:
              type: array
              nullable: true
              items:
                type: object
                properties:
                  inventoryId:
                    type: string
                    format: uuid
                    example: '123e4567-e89b-12d3-a456-426614174004'
                  inventoryItem:
                    $ref: '#/components/schemas/InventoryItem'
                  quantity:
                    type: integer
                    example: 1
                  unitPrice:
                    type: number
                    format: decimal
                    example: 16999.99
                  discount:
                    type: number
                    format: decimal
                    nullable: true
                    example: 500.00
                  total:
                    type: number
                    format: decimal
                    example: 16499.99
    CreateSaleDto:
      type: object
      required:
        - customerId
        - dealershipId
      properties:
        customerId:
          type: string
          format: uuid
          description: Customer ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        dealershipId:
          type: string
          format: uuid
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        salesPersonId:
          type: string
          format: uuid
          nullable: true
          description: Sales person ID
          example: '123e4567-e89b-12d3-a456-426614174003'
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
          description: Initial sales stage
          example: 'lead'
          default: 'lead'
        totalAmount:
          type: number
          format: decimal
          nullable: true
          description: Total sale amount
          example: 16999.99
        items:
          type: array
          nullable: true
          items:
            type: object
            required:
              - inventoryId
              - quantity
              - unitPrice
            properties:
              inventoryId:
                type: string
                format: uuid
                example: '123e4567-e89b-12d3-a456-426614174004'
              quantity:
                type: integer
                minimum: 1
                example: 1
              unitPrice:
                type: number
                format: decimal
                example: 16999.99
              discount:
                type: number
                format: decimal
                nullable: true
                example: 500.00
          description: Sale items
        notes:
          type: string
          nullable: true
          description: Sale notes
          example: 'Initial inquiry via website'
        estimatedCloseDate:
          type: string
          format: date-time
          nullable: true
          description: Estimated close date
          example: '2024-02-15T00:00:00Z'
    UpdateSaleDto:
      type: object
      properties:
        salesPersonId:
          type: string
          format: uuid
          nullable: true
          description: Sales person ID
          example: '123e4567-e89b-12d3-a456-426614174003'
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
          description: Sales stage
          example: 'negotiation'
        totalAmount:
          type: number
          format: decimal
          nullable: true
          description: Total sale amount
          example: 15999.99
        items:
          type: array
          nullable: true
          items:
            type: object
            required:
              - inventoryId
              - quantity
              - unitPrice
            properties:
              inventoryId:
                type: string
                format: uuid
                example: '123e4567-e89b-12d3-a456-426614174004'
              quantity:
                type: integer
                minimum: 1
                example: 1
              unitPrice:
                type: number
                format: decimal
                example: 16999.99
              discount:
                type: number
                format: decimal
                nullable: true
                example: 1000.00
          description: Sale items
        notes:
          type: string
          nullable: true
          description: Sale notes
          example: 'Customer negotiating price, considering trade-in'
        estimatedCloseDate:
          type: string
          format: date-time
          nullable: true
          description: Estimated close date
          example: '2024-02-10T00:00:00Z'
        actualCloseDate:
          type: string
          format: date-time
          nullable: true
          description: Actual close date (when stage is closed_won/closed_lost)
          example: '2024-02-08T16:45:00Z'
...
    PaginatedSales:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Sale'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # Service Appointment Schemas
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human readable error message
          example: 'Resource not found'
        error:
          type: string
          description: Error type
          example: 'NotFound'
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: '/api/v1/customers/invalid-id'
    ValidationError:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          description: General validation error message
          example: 'Validation failed'
        error:
          type: string
          description: Error type
          example: 'Bad Request'
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: '/api/v1/customers'
        errors:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that failed validation
                example: 'email'
              message:
                type: string
                description: Validation error message
                example: 'must be a valid email address'
              value:
                description: Invalid value that was provided
                example: 'invalid-email'
...